削除機能（修正）

+++ app/controllers/items_controller.rb
@@ -2,7 +2,7 @@ class ItemsController < ApplicationController
 
   before_action :authenticate_user!, except: [:index, :show]
   before_action :move_to_index, except: [:index, :show]
-  before_action :set_item, only: [:show, :edit]
+  before_action :set_item, except: [:index, :new, :create]
 
  def index
    @items = Item.includes(:user).order("created_at DESC")
@@ -31,7 +31,6 @@ class ItemsController < ApplicationController
  end
 
  def update
-    item = Item.find(params[:id])
     if item.update(item_params)
       redirect_to root_path
     else
@@ -40,9 +39,11 @@ class ItemsController < ApplicationController
  end
 
  def destroy
-  item  = Item.find(params[:id])
-  if item.destroy
+  if current_user.id != @item.user_id
     redirect_to root_path
+      if item.destroy
+        redirect_to root_path
+      end
   end
  end
